% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalProp.R
\name{signalProp}
\alias{signalProp}
\alias{signalProp,ANY,ANY-method}
\alias{signalProp,ANY,matrix-method}
\alias{signalProp,ANY,big.matrix-method}
\alias{signalProp,ANY,ff_matrix-method}
\alias{signalProp,numeric,missingOrNull-method}
\alias{signalProp,matrix,missingOrNull-method}
\alias{signalProp,big.matrix,missingOrNull-method}
\alias{signalProp,ff_matrix,missingOrNull-method}
\alias{signalProp,ff_array,missingOrNull-method}
\alias{signalProp,ff_vector,missingOrNull-method}
\title{Signal Proportion Estimator}
\usage{
signalProp(pval, pval_null, ...)

\S4method{signalProp}{ANY,ANY}(pval, pval_null, ...)

\S4method{signalProp}{ANY,matrix}(pval, pval_null, ..., alpha = 0.1)

\S4method{signalProp}{ANY,big.matrix}(pval, pval_null, ..., alpha = 0.1)

\S4method{signalProp}{ANY,ff_matrix}(pval, pval_null, ..., alpha = 0.1)

\S4method{signalProp}{numeric,missingOrNull}(pval, pval_null, ..., c05, c1)

\S4method{signalProp}{matrix,missingOrNull}(pval, pval_null, ...)

\S4method{signalProp}{big.matrix,missingOrNull}(pval, pval_null, ..., c05, c1)

\S4method{signalProp}{ff_matrix,missingOrNull}(pval, pval_null, ..., c05, c1)

\S4method{signalProp}{ff_array,missingOrNull}(pval, pval_null, ..., c05, c1)

\S4method{signalProp}{ff_vector,missingOrNull}(pval, pval_null, ..., c05, c1)
}
\arguments{
\item{pval}{A numeric vector object, a big.matrix object with a
single row or col, or an ff_vector, ff_array of single dimension, or
ff_matrix with a single row or col {p}. The p-values. See Details.}

\item{pval_null}{A numeric matrix object, a big.matrix object, or an
ff_matrix object of dimension {p x n}. The p-values generated 
from the null distribution. The columns correspond to the samples (n),
the rows to the signals (p). If not provided (i.e., missing or NULL),
inputs c05 and c1 must be provided. If pval_null is provided, only input 
alpha is required. See Details.}

\item{...}{Ignored. Used only to require named inputs.}

\item{alpha}{A numeric object. The significance level. The bounding sequence
is estimated as the (1-alpha)-th quantile. This input is required only if
pval_null is provided as input. See Details.}

\item{c05}{A numeric object. The bounding sequence c_{p,0.5} estimated
taking delta equal to the square root of the p-value. This input is 
required only if pval_null is not provided as input. See Details.}

\item{c1}{A numeric object. The bounding sequence c_{p,1} estimated 
taking delta equal to the p-value. This input is 
required only if pval_null is not provided as input. See Details.}
}
\value{
An S3 object of class wsiHD comprising a list object. 
  The exact contents depend on the input combination.
  The list will always include elements:
  \item{piHat}{Estimated signal proportion.}
  \item{piHat05}{Estimated signal proportion when delta is set to the
      square root of the p-value.}                 
  \item{piHat1}{Estimated signal proportion when delta is set equal
      to the p-value.}
  If the bounding sequences are estimated internally, the list will also
  include
  \item{c05}{Estimated bounding sequence when delta is set to the
      square root of the p-value.}                 
  \item{c1}{Estimated bounding sequence when delta is set equal
      to the p-value.}
}
\description{
Implements an estimator of the signal proportion in settings with 
  arbitrary covariance dependence.
}
\details{
This function is multi-use in the sense that the algorithm depends on the
  combination of inputs provided. If inputs pval, c05, and c1 are provided,
  the function uses the provided bounding sequence estimates (c05, c1) 
  to estimate the signal proportion. In contrast, if inputs pval, pval_null,
  and alpha are provided, the function uses pval_null and alpha to estimate
  the bounding sequences, which are then used to estimate the signal 
  proportion. This latter scenario is equivalent to calling 
  cSeq(pval_null, alpha) to obtain c05 and c1 and providing these as
  inputs signalProp(pval, c05, c1).

The null p-values can be provided as a numeric matrix, a big.matrix as defined
  by the bigmemory package, or as an ff_matrix as defined by the ff package.
  The latter two options allow for larger matrices. Please see the 
  documentation of these packages for details on creating objects. 

The p-values can be provided as a numeric vector, a big.matrix with a single
  column or row, or as an ff object of class (ff_vector, ff_array of 1
  dimension or ff_matrix with a single column or row).
  The latter two options allow for larger vectors.  

Further note that if both pval and pval_null are provided as input, they
  do not have to "match." For example pval can be
  a standard numeric vector with pval_null specified as a big.matrix.

If estimating the bounding sequence, note that the quantile() function of 
  base R provides 9 algorithms for estimating
  the quantile, which are based on the definitions of Hyndman and Fan (1996). 
  We have chosen the default (type = 7) here. However,
  the quantile algorithm implemented in Armadillo is type = 5 and that
  of ff is type = 1.
  Thus the results obtained using base, bigmemory, and ff objects
  containing equivalent data might differ slightly.
}
\examples{

   data(wsiData)

   # limit data to expedite example
   smp <- sample(x = 2:4089, size = 500, replace = FALSE)

   Sigma <- stats::cor(x = wsiData[,smp])

   n <- 100L
   p <- ncol(x = Sigma)

   zz <- MASS::mvrnorm(n = n, mu = rep(x = 0.0, times = p), Sigma = Sigma)
   pval_null <- {1.0 - stats::pnorm(q = abs(x = zz))}*2.0

   pval <- stats::runif(n = p)

   cseq <- cSeq(pval_null = pval_null, alpha = 0.1)
   signalProp(pval = pval, c05 = cseq$c05, c1 = cseq$c1)

   # or equivalently obtained in one step as
   signalProp(pval = pval, pval_null = pval_null, alpha = 0.1)

}
\references{
Jeng, X. J. (2021). 
    Estimating the proportion of signal variables under arbitrary
    covariance dependence. <arXiv:2102.09053>.
}
