% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsiData.R
\name{wsiData}
\alias{wsiData}
\title{Microarray Data -- Riboflavin Production in Bacillus Subtilis}
\format{
wsiData is a data.frame containing 71 participants. 
  The data.frame contains 4089 columns. The first column ($q_RIBFLC) contains
  the logarithm of the riboflavin production rate; each of the other columns
  is the expression level for a single gene.
}
\usage{
data(wsiData)
}
\description{
This dataset is available from the supplemental materials of
the given reference. 
The dataset comprises 4088 gene expression levels and the
logarithm of the riboflavin production rate for 71 individuals. This
dataset is provided for the purposes of facilitating examples.
}
\references{
Buhlmann, P., Kalisch, M. and Meier, L. (2014).
High-Dimensional statistics with a view toward applications in Biology.
Annual Review of Statistics and Its Application, 1, 255--278.
}
\keyword{datasets}
